third_party/WebKit/LayoutTests/fast/events/message-event-max-ports.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Excessively long port sequences should gracefully fail.');

shouldThrow('new MessageEvent("message", {ports:{length: 1 << 29}})');
</script>
</body>
</html>
