third_party/WebKit/LayoutTests/fast/events/message-event-max-ports-expected.txt

Excessively long port sequences should gracefully fail.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new MessageEvent("message", {ports:{length: 1 << 29}}) threw exception RangeError: Failed to construct 'MessageEvent': Array length exceeds supported limit..
PASS successfullyParsed is true

TEST COMPLETE
